+++ src/main/java/christmas/domain/discounts/BenefitDiscount.java
@@ -1,4 +1,6 @@
-package christmas.domain;
+package christmas.domain.discounts;
+
+import static christmas.constants.StringConstants.WON_FORMAT;
 
 public class BenefitDiscount implements Discount {
     private static final int BENEIFT_DISCOUNT = 25000;
@@ -19,5 +21,10 @@ public class BenefitDiscount implements Discount {
     @Override
     public int reqeustDiscountAmount() {
         return BENEIFT_DISCOUNT;
+    }
+
+    @Override
+    public String toString() {
+        return "증정 이벤트: -" + WON_FORMAT.format(BENEIFT_DISCOUNT);
     }
 }
