+++ src/main/java/christmas/domain/discounts/ChristmasDiscount.java
@@ -1,12 +1,15 @@
-package christmas.domain;
+package christmas.domain.discounts;
 
-public class ChristmasDiscount implements Discount{
+import static christmas.constants.StringConstants.WON_FORMAT;
+
+import christmas.domain.Date;
+
+public class ChristmasDiscount implements Discount {
 
     private static final int CHRISTMAS = 25;
     private static final int INITIAL_DISCOUNT = 1000;
     private static final int DAY_DISCOUNT = 100;
     private static final int FIRST_DAY = 1;
-
     private final Date date;
 
     private ChristmasDiscount(final Date date){
@@ -23,5 +26,10 @@ public class ChristmasDiscount implements Discount{
     @Override
     public int reqeustDiscountAmount() {
         return INITIAL_DISCOUNT + (date.getDifferent(FIRST_DAY)) * DAY_DISCOUNT;
+    }
+
+    @Override
+    public String toString() {
+        return "크리스마스 디데이 할인: -" + WON_FORMAT.format(INITIAL_DISCOUNT + (date.getDifferent(FIRST_DAY)) * DAY_DISCOUNT);
     }
 }
