+++ src/main/java/christmas/domain/discounts/Discount.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.discounts;
 
 public interface Discount {
     int reqeustDiscountAmount();
